+++ src/client/common/application/commands.ts
@@ -6,7 +6,10 @@
 import { CancellationToken, Position, TextDocument, Uri } from 'vscode';
 import { Commands as LSCommands } from '../../activation/languageServer/constants';
 import { Commands as DSCommands } from '../../datascience/constants';
-import { INotebook } from '../../datascience/types';
+import type { IEditCell, IInsertCell, ISaveAll, ISwapCells } from '../../datascience/interactive-common/interactiveWindowTypes';
+import type { LiveKernelModel } from '../../datascience/jupyter/kernels/types';
+import { ICell, IJupyterKernelSpec, INotebook } from '../../datascience/types';
+import type { PythonInterpreter } from '../../interpreter/contracts';
 import { CommandSource } from '../../testing/common/constants';
 import { TestFunction, TestsToRun } from '../../testing/common/types';
 import { TestDataItem, TestWorkspaceFolder } from '../../testing/types';
@@ -142,4 +145,15 @@ export interface ICommandNameArgumentTypeMapping extends ICommandNameWithoutArgu
     [DSCommands.ScrollToCell]: [string, string];
     [DSCommands.ViewJupyterOutput]: [];
     [DSCommands.SwitchJupyterKernel]: [INotebook | undefined];
+    [DSCommands.NotebookStorage_DeleteAllCells]: [Uri];
+    [DSCommands.NotebookStorage_Close]: [Uri];
+    [DSCommands.NotebookStorage_ModifyCells]: [Uri, ICell[]];
+    [DSCommands.NotebookStorage_EditCell]: [Uri, IEditCell];
+    [DSCommands.NotebookStorage_InsertCell]: [Uri, IInsertCell];
+    [DSCommands.NotebookStorage_RemoveCell]: [Uri, string];
+    [DSCommands.NotebookStorage_SwapCells]: [Uri, ISwapCells];
+    [DSCommands.NotebookStorage_Save]: [Uri, ISaveAll];
+    [DSCommands.NotebookStorage_ClearCellOutputs]: [Uri];
+    [DSCommands.NotebookStorage_SaveAs]: [Uri, Uri, ICell[]];
+    [DSCommands.NotebookStorage_UpdateVersion]: [Uri, PythonInterpreter | undefined, IJupyterKernelSpec | LiveKernelModel | undefined];
 }
